move QP return methods into new module
