+++ lib/neo4j.rb
@@ -45,6 +45,7 @@ require 'neo4j/active_rel'
 
 require 'neo4j/active_node/query_methods'
 require 'neo4j/active_node/query/query_proxy_methods'
+require 'neo4j/active_node/query/query_proxy_enumerable'
 require 'neo4j/active_node/query/query_proxy_find_in_batches'
 require 'neo4j/active_node/labels'
 require 'neo4j/active_node/id_property'
