+++ lib/neo4j/active_node/query/query_proxy.rb
@@ -2,7 +2,8 @@ module Neo4j
   module ActiveNode
     module Query
       class QueryProxy
-        include Enumerable
+
+        include Neo4j::ActiveNode::Query::QueryProxyEnumerable
         include Neo4j::ActiveNode::Query::QueryProxyMethods
         include Neo4j::ActiveNode::Query::QueryProxyFindInBatches
 
@@ -61,90 +62,6 @@ module Neo4j
           @rel_var
         end
 
-
-        # Executes the query against the database if the results are not already present in a node's association cache. This method is
-        # shared by <tt>each</tt>, <tt>each_rel</tt>, and <tt>each_with_rel</tt>.
-        # @param [String,Symbol] node The string or symbol of the node to return from the database.
-        # @param [String,Symbol] rel The string or symbol of a relationship to return from the database.
-        def enumerable_query(node, rel = nil)
-          pluck_this = rel.nil? ? [node] : [node, rel]
-          return self.pluck(*pluck_this) if @association.nil? || caller.nil?
-          cypher_string = self.to_cypher_with_params(pluck_this)
-          association_collection = caller.association_instance_get(cypher_string, @association)
-          if association_collection.nil?
-            association_collection = self.pluck(*pluck_this)
-            caller.association_instance_set(cypher_string, association_collection, @association) unless association_collection.empty?
-          end
-          association_collection
-        end
-
-        # Just like every other <tt>each</tt> but it allows for optional params to support the versions that also return relationships.
-        # The <tt>return_node</tt> and <tt>return_rel</tt> params are typically used by those other methods but there's nothing stopping you from
-        # using `your_node.each(true, true)` instead of `your_node.each_with_rel`.
-        # @return [Enumerable] An enumerable containing some combination of nodes and rels.
-        def each(return_node = true, return_rel = false)
-          if return_node && return_rel
-            enumerable_query(identity, @rel_var).each { |obj, rel| yield obj, rel }
-          else
-            pluck_this = !return_rel ? identity : @rel_var
-            enumerable_query(pluck_this).each { |obj| yield obj }
-          end
-        end
-
-        # When called at the end of a QueryProxy chain, it will return the resultant relationship objects intead of nodes.
-        # For example, to return the relationship between a given student and their lessons:
-        #   student.lessons.each_rel do |rel|
-        # @return [Enumerable] An enumerable containing any number of applicable relationship objects.
-        def each_rel(&block)
-          block_given? ? each(false, true, &block) : to_enum(:each, false, true)
-        end
-
-        # When called at the end of a QueryProxy chain, it will return the nodes and relationships of the last link.
-        # For example, to return a lesson and each relationship to a given student:
-        #   student.lessons.each_with_rel do |lesson, rel|
-        def each_with_rel(&block)
-          block_given? ? each(true, true, &block) : to_enum(:each, true, true)
-        end
-
-        # Used as part of `dependent: :destroy` and may not have any utility otherwise.
-        # It keeps track of the node responsible for a cascading `destroy` process.
-        # @param [#dependent_children] caller The node that called this method. Typically, we would use QueryProxy's `caller` method
-        # but this is not always available, so we require it explicitly.
-        def each_for_destruction(owning_node)
-          target = owning_node.called_by || owning_node
-          enumerable_query(identity).each do |obj|
-            # Cypher can return nil objects, check for empty results
-            next if !obj || target.dependent_children.include?(obj)
-            obj.called_by = target
-            target.dependent_children << obj
-            yield obj
-          end
-        end
-
-        # Does exactly what you would hope. Without it, comparing `bobby.lessons == sandy.lessons` would evaluate to false because it
-        # would be comparing the QueryProxy objects, not the lessons themselves.
-        def ==(other)
-          self.to_a == other
-        end
-
-        METHODS = %w(where rel_where order skip limit)
-
-        METHODS.each do |method|
-          module_eval(%{
-            def #{method}(*args)
-              build_deeper_query_proxy(:#{method}, args)
-            end}, __FILE__, __LINE__)
-        end
-        # Since there is a rel_where method, it seems only natural for there to be node_where
-        alias_method :node_where, :where
-        alias_method :offset, :skip
-        alias_method :order_by, :order
-
-        # For getting variables which have been defined as part of the association chain
-        def pluck(*args)
-          self.query.pluck(*args)
-        end
-
         def params(params)
           self.dup.tap do |new_query|
             new_query._add_params(params)
@@ -190,7 +107,18 @@ module Neo4j
           @model.current_scope = previous
         end
 
+        METHODS = %w(where rel_where order skip limit)
 
+        METHODS.each do |method|
+          module_eval(%{
+            def #{method}(*args)
+              build_deeper_query_proxy(:#{method}, args)
+            end}, __FILE__, __LINE__)
+        end
+        # Since there is a rel_where method, it seems only natural for there to be node_where
+        alias_method :node_where, :where
+        alias_method :offset, :skip
+        alias_method :order_by, :order
 
         # Cypher string for the QueryProxy's query. This will not include params. For the full output, see <tt>to_cypher_with_params</tt>.
         def to_cypher
